Fix warnings

+++ Razer/ChromaSDKPluginTypes.cpp
@@ -8,8 +8,8 @@ FChromaSDKColors::FChromaSDKColors()
 FChromaSDKColors::FChromaSDKColors(const FChromaSDKColors& copy)
 {
 	const std::vector<COLORREF>& colors = copy.Colors;
-	int size = colors.size();
-	for (int index = 0; index < size; ++index)
+	size_t size = colors.size();
+	for (size_t index = 0; index < size; ++index)
 	{
 		const COLORREF& color = colors[index];
 		Colors.push_back(color);
@@ -23,8 +23,8 @@ FChromaSDKColorFrame1D::FChromaSDKColorFrame1D()
 FChromaSDKColorFrame1D::FChromaSDKColorFrame1D(const FChromaSDKColorFrame1D& copy)
 {
 	const std::vector<COLORREF>& colors = copy.Colors;
-	int size = colors.size();
-	for (int index = 0; index < size; ++index)
+	size_t size = colors.size();
+	for (size_t index = 0; index < size; ++index)
 	{
 		const COLORREF& color = colors[index];
 		Colors.push_back(color);
@@ -39,8 +39,8 @@ FChromaSDKColorFrame2D::FChromaSDKColorFrame2D()
 FChromaSDKColorFrame2D::FChromaSDKColorFrame2D(const FChromaSDKColorFrame2D& copy)
 {
 	const std::vector<FChromaSDKColors>& colors = copy.Colors;
-	int size = colors.size();
-	for (int index = 0; index < size; ++index)
+	size_t size = colors.size();
+	for (size_t index = 0; index < size; ++index)
 	{
 		const FChromaSDKColors& color = colors[index];
 		Colors.push_back(color);
